Refactor bandwidth parsing module
